Make website redirect to https://
It's done through flask, so that every page will always be secure. I got
the certificate so it's gucci now.

Also updates flip and about.html to use the new secure URLs

Actually it uses // urls so that it will match whatever it's using, but
flask will redirect to https anyway.